Lynxchan refactor: move sendPost and deletePost methods to Endchan module
Added internal class for captcha with ID
