+++ src/nya/miku/wishmaster/api/AbstractLynxChanModule.java
@@ -23,7 +23,6 @@ import android.content.res.Resources;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -52,9 +51,7 @@ import nya.miku.wishmaster.api.models.AttachmentModel;
 import nya.miku.wishmaster.api.models.BadgeIconModel;
 import nya.miku.wishmaster.api.models.BoardModel;
 import nya.miku.wishmaster.api.models.CaptchaModel;
-import nya.miku.wishmaster.api.models.DeletePostModel;
 import nya.miku.wishmaster.api.models.PostModel;
-import nya.miku.wishmaster.api.models.SendPostModel;
 import nya.miku.wishmaster.api.models.SimpleBoardModel;
 import nya.miku.wishmaster.api.models.ThreadModel;
 import nya.miku.wishmaster.api.models.UrlPageModel;
@@ -63,16 +60,10 @@ import nya.miku.wishmaster.api.util.CryptoUtils;
 import nya.miku.wishmaster.api.util.RegexUtils;
 import nya.miku.wishmaster.api.util.UrlPathUtils;
 import nya.miku.wishmaster.api.util.WakabaUtils;
-import nya.miku.wishmaster.common.IOUtils;
 import nya.miku.wishmaster.common.Logger;
-import nya.miku.wishmaster.http.JSONEntry;
-import nya.miku.wishmaster.http.interactive.SimpleCaptchaException;
 import nya.miku.wishmaster.http.streamer.HttpRequestModel;
 import nya.miku.wishmaster.http.streamer.HttpResponseModel;
 import nya.miku.wishmaster.http.streamer.HttpStreamer;
-import nya.miku.wishmaster.lib.MimeTypes;
-import nya.miku.wishmaster.lib.base64.Base64;
-import nya.miku.wishmaster.lib.base64.Base64OutputStream;
 import nya.miku.wishmaster.lib.org_json.JSONArray;
 import nya.miku.wishmaster.lib.org_json.JSONObject;
 
@@ -89,9 +80,7 @@ public abstract class AbstractLynxChanModule extends AbstractWakabaModule {
     private static final Pattern GREEN_TEXT_MARK_PATTERN = Pattern.compile("<span class=\"greenText\">(.*?)</span>");
     private static final Pattern REPLY_NUMBER_PATTERN = Pattern.compile("&gt&gt(\\d+)");
     protected Map<String, BoardModel> boardsMap = null;
-    private Map<String, ArrayList<String>> flagsMap = null;
-    private static String lastCaptchaId;
-    private static String lastCaptchaAnswer;
+    protected Map<String, ArrayList<String>> flagsMap = null;
     
     public AbstractLynxChanModule(SharedPreferences preferences, Resources resources) {
         super(preferences, resources);
@@ -410,7 +399,7 @@ public abstract class AbstractLynxChanModule extends AbstractWakabaModule {
         return super.fixRelativeUrl(url);
     }
 
-    protected CaptchaModel downloadCaptcha(String captchaUrl, ProgressListener listener, CancellableTask task) throws Exception {
+    protected ExtendedCaptchaModel downloadCaptcha(String captchaUrl, ProgressListener listener, CancellableTask task) throws Exception {
         Bitmap captchaBitmap = null;
         HttpRequestModel requestModel = HttpRequestModel.DEFAULT_GET;
         HttpResponseModel responseModel = HttpStreamer.getInstance().getFromUrl(captchaUrl, requestModel, httpClient, listener, task);
@@ -433,10 +422,10 @@ public abstract class AbstractLynxChanModule extends AbstractWakabaModule {
         } finally {
             responseModel.release();
         }
-        CaptchaModel captchaModel = new CaptchaModel();
+        ExtendedCaptchaModel captchaModel = new ExtendedCaptchaModel();
         captchaModel.type = CaptchaModel.TYPE_NORMAL;
         captchaModel.bitmap = captchaBitmap;
-        lastCaptchaId = captchaId;
+        captchaModel.captchaID = captchaId;
         return captchaModel;
     }
 
@@ -446,7 +435,7 @@ public abstract class AbstractLynxChanModule extends AbstractWakabaModule {
         return downloadCaptcha(captchaUrl, listener, task);
     }
 
-    private String computeFileMD5(File file) throws NoSuchAlgorithmException, IOException {
+    public static String computeFileMD5(File file) throws NoSuchAlgorithmException, IOException {
         MessageDigest messageDigest = MessageDigest.getInstance("MD5");
         messageDigest.reset();
 
@@ -469,149 +458,7 @@ public abstract class AbstractLynxChanModule extends AbstractWakabaModule {
         return md5Hex;
     }
 
-    private String checkFileIdentifier(File file, String mime, ProgressListener listener, CancellableTask task) {
-        String hash;
-        try {
-            hash = computeFileMD5(file);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-        String identifier = hash + "-" + mime.replace("/", "");
-        String url = getUsingUrl() + "checkFileIdentifier.js?identifier=" + identifier;
-        String response = "";
-        try {
-            response = HttpStreamer.getInstance().getStringFromUrl(url, null, httpClient, listener, task, false);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-        if (response.contains("true")) return hash;
-        return null;
+    public class ExtendedCaptchaModel extends CaptchaModel {
+        public String captchaID = "";
     }
-
-    private String base64EncodeFile(File file) throws IOException {
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
-        Base64OutputStream b64os = new Base64OutputStream(os, Base64.NO_WRAP);
-        FileInputStream fis = new FileInputStream(file);
-        IOUtils.copyStream(fis, b64os);
-        return os.toString();
-    }
-
-    public String sendPost(SendPostModel model, ProgressListener listener, CancellableTask task) throws Exception {
-        String url = getUsingUrl() + ".api/" + (model.threadNumber == null ? "newThread" : "replyThread");
-
-        JSONObject jsonPayload = new JSONObject();
-        JSONObject jsonParameters = new JSONObject();
-        jsonPayload.put("captchaId", lastCaptchaId);
-        jsonParameters.put("name", model.name);
-        jsonParameters.put("password", model.password);
-        jsonParameters.put("subject", model.subject);
-        jsonParameters.put("message", model.comment);
-        jsonParameters.put("boardUri", model.boardName);
-        jsonParameters.put("email", model.sage ? "sage" : model.email);
-        if (model.threadNumber != null)
-            jsonParameters.put("threadId", model.threadNumber);
-        if (model.captchaAnswer != null && model.captchaAnswer.length() > 0)
-            jsonParameters.put("captcha", model.captchaAnswer);
-        if (model.icon > 0)
-            jsonParameters.put("flag", flagsMap.get(model.boardName).get(model.icon - 1));
-        if (model.attachments != null && model.attachments.length > 0) {
-            JSONArray files = new JSONArray();
-            for (int i = 0; i < model.attachments.length; ++i) {
-                String name = model.attachments[i].getName();
-                String mime = MimeTypes.forExtension(name.substring(name.lastIndexOf('.') + 1), "");
-                String md5 = checkFileIdentifier(model.attachments[i], mime, listener, task);
-                JSONObject file = new JSONObject();
-                file.put("name", name);
-                if (md5 != null) {
-                    file.put("md5", md5);
-                    file.put("mime", mime);
-                } else {
-                    file.put("content", "data:" + mime + ";base64," + base64EncodeFile(model.attachments[i]));
-                }
-                file.put("spoiler", false);
-                files.put(file);
-            }
-            jsonParameters.put("spoiler", model.custommark);
-            jsonParameters.put("files", files);
-        }
-        jsonPayload.put("parameters", jsonParameters);
-        JSONEntry payload = new JSONEntry(jsonPayload);
-        HttpRequestModel request = HttpRequestModel.builder().setPOST(payload).setNoRedirect(true).build();
-        String response = HttpStreamer.getInstance().getStringFromUrl(url, request, httpClient, null, task, true);
-        lastCaptchaId = null;
-        JSONObject result = new JSONObject(response);
-        String status = result.optString("status");
-        if ("ok".equals(status)) {
-            UrlPageModel urlPageModel = new UrlPageModel();
-            urlPageModel.type = UrlPageModel.TYPE_THREADPAGE;
-            urlPageModel.chanName = getChanName();
-            urlPageModel.boardName = model.boardName;
-            urlPageModel.threadNumber = model.threadNumber;
-            if (model.threadNumber == null) {
-                urlPageModel.threadNumber = Integer.toString(result.optInt("data"));
-            } else {
-                urlPageModel.postNumber = Integer.toString(result.optInt("data"));
-            }
-            return buildUrl(urlPageModel);
-        } else if (status.contains("error") || status.contains("blank")) {
-            String errorMessage = result.optString("data");
-            if (errorMessage.length() > 0) {
-                throw new Exception(errorMessage);
-            }
-        }
-        throw new Exception("Unknown Error");
-    }
-
-    @Override
-    public String deletePost(DeletePostModel model, final ProgressListener listener, final CancellableTask task) throws Exception {
-        String url = getUsingUrl() + ".api/" + "deleteContent";
-        
-        if (lastCaptchaAnswer == null) {
-            throw new SimpleCaptchaException() {
-                private static final long serialVersionUID = 1L;
-                @Override
-                protected Bitmap getNewCaptcha() throws Exception {
-                    return AbstractLynxChanModule.this.getNewCaptcha("", "", listener, task).bitmap;
-                }
-                @Override
-                protected void storeResponse(String response) {
-                    lastCaptchaAnswer = response;
-                }
-            };
-        }
-        
-        JSONObject jsonPayload = new JSONObject();
-        JSONObject jsonParameters = new JSONObject();
-        jsonPayload.put("captchaId", lastCaptchaId);
-        jsonParameters.put("password", model.password);
-        if (lastCaptchaAnswer != null && lastCaptchaAnswer.length() > 0)
-            jsonParameters.put("captcha", lastCaptchaAnswer);
-        jsonParameters.put("deleteMedia", true);
-        if (model.onlyFiles) {
-            jsonParameters.put("deleteUploads", true);
-        }
-        JSONArray jsonArray = new JSONArray();
-        JSONObject post = new JSONObject();
-        post.put("board", model.boardName);
-        post.put("thread", model.threadNumber);
-        if (!model.postNumber.equals(model.threadNumber)) post.put("post", model.postNumber);
-        jsonParameters.put("postings", jsonArray);
-        jsonPayload.put("parameters", jsonParameters);
-        JSONEntry payload = new JSONEntry(jsonPayload);
-        HttpRequestModel request = HttpRequestModel.builder().setPOST(payload).setNoRedirect(true).build();
-        String response = HttpStreamer.getInstance().getStringFromUrl(url, request, httpClient, null, task, true);
-        lastCaptchaId = null;
-        lastCaptchaAnswer = null;
-        JSONObject result = new JSONObject(response);
-        if (result.optString("status").equals("error")) {
-            String errorMessage = result.optString("data");
-            if (errorMessage.length() > 0) {
-                throw new Exception(errorMessage);
-            }
-        }
-        return null;
-    }
-    
 }
